Переносит функцию создания миниатюры в отдельный модуль create-thumbnail.js
