+++ js/sorting.js
@@ -1,6 +1,7 @@
 import {openBigPicture} from'./viewer-image.js';
 import {container} from './thumbnail-rendering.js';
-import {createPhotoThumbnails, createThumbnail} from './thumbnail-rendering.js';
+import {createPhotoThumbnails} from './thumbnail-rendering.js';
+import {createThumbnail} from './cteate-tumbnail.js'
 import {configFilter} from './filter.js';
 import {showErrorMessage} from './error-message.js';
